change type data to uint16

+++ src/Config.h
@@ -4,9 +4,9 @@
 #define I2C_PIN_SCL A5
 
 namespace SystemTiming {
-constexpr uint8_t DEBOUNCE_MS = 100;
-constexpr uint8_t LONGPRESS_MS = 1000;
-constexpr uint8_t BLINK_MS = 500;
+constexpr uint16_t DEBOUNCE_MS = 100;
+constexpr uint16_t LONGPRESS_MS = 1000;
+constexpr uint16_t BLINK_MS = 500;
 }
 
 #define DS3231_ADDRESS 0x68
